Refactor: 컴포넌트 분리 및 중복코드 삭제
